requirements.txt
# Scraping
selenium
webdriver-manager

#cleaning
pandas
pyspark

# AI agents and Streamlit
streamlit
pandas
matplotlib
seaborn
Pillow==10.4.0
together==1.3.10
e2b==1.0.5
e2b-code-interpreter==1.0.3

#gcp
google-cloud-storage
google-cloud-sdk
